flask-backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.22
python-dotenv==0.19.0
mysql-connector-python==8.0.26
pyspark==3.1.2
Werkzeug==2.0.1
pymysql==1.1.1
markupsafe==2.0.1
